test: extract
